% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_data}
\alias{age_data}
\title{Summary of Irish age data by cohort}
\format{\describe{
\item{AgeCohort}{The age cohort}
\item{Population}{The population}
}}
\usage{
age_data
}
\description{
Summary of Irish age data by cohort
}
\keyword{datasets}
