% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirR.R
\name{create_seir}
\alias{create_seir}
\title{Creates an object to facilitate running the simulation}
\usage{
create_seir()
}
\value{
An S3 object of class seir
}
\description{
\code{create_seir} returns an seir object to the calling program.
This object contains a tibble with default parameter, any of which
can be modified before the model is rum
}
\examples{
\dontrun{
mod <- create_seir()
}
}
