% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_seirR.R
\name{create_seir_p}
\alias{create_seir_p}
\title{Creates an object to facilitate running the simulation}
\usage{
create_seir_p(model_offset = 0)
}
\arguments{
\item{model_offset}{the offset for the model (subtracted from start time)}
}
\value{
An S3 object of class seir
}
\description{
\code{create_seir_p} returns a population seir object to the calling program.
This object contains a tibble with default parameter, any of which
can be modified before the model is run
}
\examples{
\dontrun{
mod <- create_seir_p()
}
}
