% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirR.R
\name{explain}
\alias{explain}
\title{Explains the current model assumptions}
\usage{
explain(o)
}
\arguments{
\item{o}{is the seir S3 object}
}
\description{
\code{explain()} summarises the model
}
\details{
As it's a generic function, this call is dispatched to run.seir
}
