% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{get_biological_params_03}
\alias{get_biological_params_03}
\title{Gets the biological params for the initial params tibble}
\usage{
get_biological_params_03(tb)
}
\arguments{
\item{tb}{is input seir object}
}
\value{
tibble of class seir
}
\description{
\code{get_biological_params_03} returns resource params
}
