% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_get_constants_seir.R
\name{get_constants_seir}
\alias{get_constants_seir}
\title{Creates the vector of constants that can be used to set params}
\usage{
get_constants_seir()
}
\value{
A vector of constants and their initial conditions
}
\description{
Creates the vector of constants that can be used to set params
}
