% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_get_curve_analysis.R
\name{get_curve_analysis}
\alias{get_curve_analysis}
\title{Gets the behaviour modes of a variable}
\usage{
get_curve_analysis(out, indicators)
}
\arguments{
\item{out}{is the output simulation tibble}

\item{indicators}{is the list of vars to do the analysis o}
}
\value{
A tidy data set with behaviour modes
}
\description{
\code{get_curve_analysis()} returns behaviour modes for a time series
}
