% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_setup.R
\name{get_distancing_params_05}
\alias{get_distancing_params_05}
\title{Gets the physical distancing params for the initial params tibble}
\usage{
get_distancing_params_05(tb)
}
\arguments{
\item{tb}{is input seir object}
}
\value{
tibble of class seir
}
\description{
\code{get_distancing_params_05} returns resource params
}
