% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirR.R
\name{get_flow_params_04}
\alias{get_flow_params_04}
\title{Gets the flow params for the initial params tibble}
\usage{
get_flow_params_04(tb)
}
\arguments{
\item{tb}{is input seir object}
}
\value{
tibble of class seir
}
\description{
\code{get_flow_params_04} returns resource params
}
