% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_api.R
\name{get_health_system_params_06}
\alias{get_health_system_params_06}
\title{Gets the health system params for the initial params tibble}
\usage{
get_health_system_params_06(tb)
}
\arguments{
\item{tb}{is input seir object}
}
\value{
tibble of class seir
}
\description{
\code{get_health_system_params_06} returns resource params
}
