% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_initial_conditions_01}
\alias{get_initial_conditions_01}
\title{Gets the transmission params for the initial params tibble}
\usage{
get_initial_conditions_01(tb)
}
\arguments{
\item{tb}{is input seir object}
}
\value{
tibble of class seir
}
\description{
\code{get_initial_conditions_01} returns resource params
}
