% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_seirR.R
\name{get_models}
\alias{get_models}
\title{Creates an object to facilitate running the simulation}
\usage{
get_models()
}
\value{
tibble with model list
}
\description{
\code{get_models} returns a list of models
}
