% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{get_param}
\alias{get_param}
\title{Gets a parameter value}
\usage{
get_param(o, p, isString = F)
}
\arguments{
\item{o}{is the seir S3 object}

\item{p}{is the parameter name}
}
\value{
The parameter value
}
\description{
\code{get_param()} runs the simulation, based on the parameter set
}
\details{
As it's a generic function, this call is dispatched to run.seir
}
