% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_api.R
\name{get_resource_params_07}
\alias{get_resource_params_07}
\title{Gets the resource parameters for the initial params tibble}
\usage{
get_resource_params_07(tb)
}
\arguments{
\item{tb}{is input seir object}
}
\value{
tibble of class seir
}
\description{
\code{get_resource_params_07} returns resource params
}
