% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{get_transmission_params_02}
\alias{get_transmission_params_02}
\title{Gets the transmission params for the initial params tibble}
\usage{
get_transmission_params_02(tb)
}
\arguments{
\item{tb}{is input seir object}
}
\value{
tibble of class seir
}
\description{
\code{get_transmission_params_02} returns resource params
}
