% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_seir_model.R
\name{init_compartments}
\alias{init_compartments}
\title{Creates the initial state vector for ODE}
\usage{
init_compartments()
}
\value{
A vector of compartments and their initial conditions
}
\description{
Creates the initial state vector for ODE
}
