% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_compartments_seir_p.R
\name{init_compartments_p}
\alias{init_compartments_p}
\title{Creates the initial state vector for population ODE}
\usage{
init_compartments_p()
}
\value{
A vector of compartments and their initial conditions
}
\description{
Creates the initial state vector for population ODE
}
