% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_p_02_init_compartments_seir_p.R
\name{init_compartments_p_xmile}
\alias{init_compartments_p_xmile}
\title{Creates the initial state vector for population ODE}
\usage{
init_compartments_p_xmile()
}
\value{
A vector of compartments and their initial conditions
}
\description{
Creates the initial state vector for population ODE
}
