% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_seir_p_02_init_compartments.R
\name{init_seir_p_compartments}
\alias{init_seir_p_compartments}
\title{Creates the initial state vector for population ODE}
\usage{
init_seir_p_compartments()
}
\value{
A vector of compartments and their initial conditions
}
\description{
Creates the initial state vector for population ODE
}
