% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_seirR.R
\name{is_model}
\alias{is_model}
\title{Creates an object to facilitate running the simulation}
\usage{
is_model(mod)
}
\arguments{
\item{mod}{is the S3 model class}
}
\value{
logical value
}
\description{
\code{is_model} checks to see if an object is an seir S3 class
}
