% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_model_seir_p.R
\name{model_seir_p}
\alias{model_seir_p}
\title{The set of ODEs to solve for the model}
\usage{
model_seir_p(time, stocks, auxs)
}
\arguments{
\item{time}{is the current simulation time}

\item{stocks}{is the vector of model compartments}

\item{auxs}{vector of auxiliaries (alway NULL) as these are passed via \code{sim_state} environment}
}
\value{
list of new compartment values, and other relevant variables
}
\description{
\code{seir_model_p} Solves the population model for each timestep, is called by \code{ode}
}
