% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_seir_pf_01_model.R
\name{model_seir_pf}
\alias{model_seir_pf}
\title{The set of ODEs to solve for the model}
\usage{
model_seir_pf(time, stocks, auxs)
}
\arguments{
\item{time}{is the current simulation time}

\item{stocks}{is the vector of model compartments}

\item{auxs}{vector of auxiliaries (alway NULL) as these are passed via sim_state environment}
}
\value{
list of new compartment values, and other relevant variables
}
\description{
model_seir_p Solves the population model for each timestep, is called by ode
}
