% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_seirR.R
\name{params}
\alias{params}
\title{Gets a parameter value}
\usage{
params(o)
}
\arguments{
\item{o}{is the seir S3 object}
}
\value{
The parameter value
}
\description{
\code{params()} returns the parameters
}
\details{
As it's a generic function, this call is dispatched to run.seir
}
