% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_run_model_seir.R
\name{run_model_seir}
\alias{run_model_seir}
\title{Prepares all variables and calls \code{deSolve::ode()}}
\usage{
run_model_seir(mod_object, DT = 0.125, return_all = F)
}
\arguments{
\item{mod_object}{are the simulation parameters}

\item{DT}{is the simulation time step (Euler)}

\item{return_all}{a flag used to decide how many observations to return}
}
\value{
A tibble of simulation results
}
\description{
\code{run_seir_model()} runs the simulation, based on the parameter set
}
\details{
As it's a generic function, this call is dispatched to run.seir
}
