% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_model.R
\name{seir_model}
\alias{seir_model}
\title{The set of ODEs to solve for the model}
\usage{
seir_model(time, stocks, auxs)
}
\arguments{
\item{time}{is the current simulation time}

\item{stocks}{is the vector of model compartments}

\item{auxs}{vector of auxiliaries (alway NULL) as these are passed via sim_state environment}
}
\value{
list of new compartment values, and other relevant variables
}
\description{
\code{seir_model} Solves the model for each timestep, is called by \code{ode}
}
