% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_api.R
\name{set_param}
\alias{set_param}
\title{Sets a parameter}
\usage{
set_param(o, p, v)
}
\arguments{
\item{o}{is the seir S3 object}

\item{p}{is the parameter name}

\item{v}{is the new value for the parameter}
}
\value{
A tibble of simulation results
}
\description{
\code{run()} runs the simulation, based on the parameter set
}
\details{
As it's a generic function, this call is dispatched to run.seir
}
\examples{
\dontrun{
m <- create_seir()
o <- run(m)
}
}
