% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_seir_a_04_setup.R
\name{setup_seir_a_parameters}
\alias{setup_seir_a_parameters}
\title{Populates the parameters tibble for seir_p 
NOTE: Edit this file to clarify parameter types}
\usage{
setup_seir_a_parameters()
}
\description{
Populates the parameters tibble for seir_p 
NOTE: Edit this file to clarify parameter types
}
