% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_seirR.R
\name{summary.seir}
\alias{summary.seir}
\title{Explains the current model assumptions}
\usage{
\method{summary}{seir}(object, ...)
}
\arguments{
\item{object}{is the seir S3 object}

\item{...}{are additional params}
}
\description{
\code{summary()} summarises the model
}
\details{
As it's a generic function, this call is dispatched to run.seir
}
