% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_unmitigated}
\alias{test_unmitigated}
\title{Summary of test data set for unmitigated scenario}
\format{\describe{
\item{Time (Day}{The simulation time}
\item{Susceptible}{Total suceptible from Vensim run}
\item{Total Exposed}{Total exposed from Vensim run}
\item{Total Infectious}{Total infectious from Vensim run}
\item{Total Removed}{Total removed from Vensim run}
\item{Total in Hospital}{Total in hospital from Vensim run}
\item{CheckSum Population}{Checking the population remains constant}
}}
\usage{
test_unmitigated
}
\description{
Summary of test data set for unmitigated scenario
}
\keyword{datasets}
