% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmile_04_get_constants_seir_p.R
\name{xmile_get_constants}
\alias{xmile_get_constants}
\title{Creates the vector of constants that can be used to set params}
\usage{
xmile_get_constants()
}
\value{
A vector of constants and their initial conditions
}
\description{
Creates the vector of constants that can be used to set params
}
