% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_p_05_setup_seir_p.R
\name{xmile_setup_parameters}
\alias{xmile_setup_parameters}
\title{Populates the parameters tibble for seir_p 
NOTE: Edit this file to clarify parameter types}
\usage{
xmile_setup_parameters()
}
\description{
Populates the parameters tibble for seir_p 
NOTE: Edit this file to clarify parameter types
}
