% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_world_data}
\alias{get_world_data}
\title{Gets the latest world covid data}
\usage{
get_world_data()
}
\value{
tibble of class seir
}
\description{
\code{get_world_data} creates a tibble in the data_env
}
