% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_seir_model.R
\name{run_seir_model}
\alias{run_seir_model}
\title{Prepares all variables and calls \code{deSolve::ode()}}
\usage{
run_seir_model(params, start, finish, DT, return_all = F, offset = 0)
}
\arguments{
\item{params}{are the simulation parameters}

\item{start}{is thesimulation start time}

\item{finish}{is the simulation finish time}

\item{DT}{is the simulation time step (Euler)}

\item{return_all}{a flag used to decide how many observations to return}

\item{offset}{is the simulation offset time}
}
\value{
A tibble of simulation results
}
\description{
\code{run_seir_model()} runs the simulation, based on the parameter set
}
\details{
As it's a generic function, this call is dispatched to run.seir
}
