% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirR.R
\name{set_param}
\alias{set_param}
\title{Sets a parameter value}
\usage{
set_param(o, p, v, isString = F, checkStr = F)
}
\arguments{
\item{o}{is the seir S3 object}

\item{p}{is the parameter name}

\item{v}{is the new value for the parameter}

\item{isString}{to indicate if its a string param.}

\item{checkStr}{check the match}
}
\value{
A tibble of simulation results
}
\description{
\code{set_param()} runs the simulation, based on the parameter set
}
\details{
As it's a generic function, this call is dispatched to run.seir
}
\examples{
\dontrun{
m <- create_seir()
o <- run(m)
}
}
